Pull CountSpectrum Class Out & Delete Useless File
1. 把計算頻譜的class拉出來成一個java檔
2. 刪掉沒用的java檔
